Move FileSystemUtils into FileSystem
Tests were integrated into FileSystemIntegrationTest.
